Don't force alphanumeric attribute option sorting
Magento allows the administrator to customise the sorting of attribute options using the position column. MVentory was preventing this being used as it overloaded the attribute model and forced its options to be sorted alphanumerically.

This affected the front-end display of attributes in layered navigation. MVentory should either respect the Magento sorting, or only inject this behaviour in its own functions.